+++ apps/frontend/src/app/Solver/index.d.ts
@@ -68,17 +68,6 @@ export interface Interim {
   skipped: number
 }
 
-export interface WorkerRecvMessage<T = any> {
-  command: 'workerRecvMessage'
-  from: number
-  data: T
-}
-export interface WorkerSendMessage<T = any> {
-  messageType: 'send'
-  to?: number | 'all'
-  data: T
-}
-
 export type MessageData = Threshold
 export interface Threshold {
   dataType: 'threshold'
